pyproject.toml
[project]
name = "hr-breaker"
version = "0.1.1"
description = "Resume optimization tool for job postings"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Boris Tseitlin"}, {name = "Claude Code"}]
license = {text = "MIT"}
keywords = ["resume", "job", "optimization", "ats", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "streamlit>=1.30",
    "pydantic>=2.0",
    "pydantic-ai-slim[openai]==1.48.0",
    "httpx",
    "beautifulsoup4",
    "scikit-learn>=1.0",
    "python-dotenv",
    "click>=8.0",
    "pymupdf>=1.24",
    "weasyprint>=60.0",
    "jinja2>=3.1",
    "nest-asyncio>=1.6.0",
    "playwright>=1.40",
    "watchdog>=6.0.0",
]

[project.scripts]
hr-breaker = "hr_breaker.cli:cli"

[project.urls]
Homepage = "https://github.com/btseytlin/hr-breaker"
Repository = "https://github.com/btseytlin/hr-breaker"
Issues = "https://github.com/btseytlin/hr-breaker/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hr_breaker"]

[tool.hatch.build.targets.wheel.force-include]
"templates" = "hr_breaker/templates"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
